refactor: AnswerService 핸들러 패턴 적용 및 타입별 핸들러 구현 (#63)
